+++ CHANGES.md
@@ -19,3 +19,5 @@
 - (2021-10-28) Added base data_handler
 
 - (2021-10-28) Merge to develop
+
+- (2021-10-28) Added base structure rest_flask